Finish initial scheduler implementation

+++ core/src/emulator.rs
@@ -13,7 +13,7 @@ use crate::io::joypad::*;
 
 /// A DMG runs at `4.194304 MHz` with a Vsync of `59.7275 Hz`, so that would be
 /// `4194304 / 59.7275 = 70224 cycles/frame`
-pub const CYCLES_PER_FRAME: u32 = 70224;
+pub const CYCLES_PER_FRAME: u64 = 70224;
 
 pub struct Emulator {
     cpu: CPU<Memory>,
